wildlife-ai/requirements.txt
# ============================================================
# Wildlife AI — Python Backend Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# For GPU:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# ============================================================

# ── Web Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9       # Required for UploadFile in FastAPI
httpx==0.27.0                 # Async HTTP calls to Node service

# ── AI / Deep Learning ─────────────────────────────────────
torch==2.3.0
torchvision==0.18.0
ultralytics==8.2.0            # YOLOv8

# ── Image Processing ───────────────────────────────────────
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# ── Cloud Storage ──────────────────────────────────────────
cloudinary==1.40.0            # Cloudinary Python SDK

# ── Generative AI ──────────────────────────────────────────
anthropic==0.28.0             # Claude AI
openai==1.30.0                # GPT (optional fallback)

# ── Config ─────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Dev / Testing ──────────────────────────────────────────
pytest==8.2.1
